+++ super_editor/lib/src/default_editor/blockquote.dart
@@ -5,10 +5,10 @@ import 'package:super_editor/src/core/edit_context.dart';
 import 'package:super_editor/src/default_editor/attributions.dart';
 import 'package:super_editor/src/infrastructure/_logging.dart';
 import 'package:super_editor/src/infrastructure/attributed_text_styles.dart';
+import 'package:super_editor/src/infrastructure/keyboard.dart';
 
 import '../core/document.dart';
 import '../core/document_editor.dart';
-import 'document_input_keyboard.dart';
 import 'layout_single_column/layout_single_column.dart';
 import 'paragraph.dart';
 import 'text.dart';
@@ -199,23 +199,22 @@ class BlockquoteComponent extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return IgnorePointer(
-      child: Container(
-        padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
-        decoration: BoxDecoration(
-          borderRadius: borderRadius,
-          color: backgroundColor,
-        ),
-        child: TextComponent(
-          key: textKey,
-          text: text,
-          textStyleBuilder: styleBuilder,
-          textSelection: textSelection,
-          selectionColor: selectionColor,
-          highlightWhenEmpty: highlightWhenEmpty,
-          showDebugPaint: showDebugPaint,
-        ),
-      )
-    );
+        child: Container(
+      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
+      decoration: BoxDecoration(
+        borderRadius: borderRadius,
+        color: backgroundColor,
+      ),
+      child: TextComponent(
+        key: textKey,
+        text: text,
+        textStyleBuilder: styleBuilder,
+        textSelection: textSelection,
+        selectionColor: selectionColor,
+        highlightWhenEmpty: highlightWhenEmpty,
+        showDebugPaint: showDebugPaint,
+      ),
+    ));
   }
 }
