+++ super_editor/lib/src/default_editor/list_items.dart
@@ -4,6 +4,7 @@ import 'package:flutter/services.dart';
 import 'package:super_editor/src/core/edit_context.dart';
 import 'package:super_editor/src/infrastructure/_logging.dart';
 import 'package:super_editor/src/infrastructure/attributed_text_styles.dart';
+import 'package:super_editor/src/infrastructure/keyboard.dart';
 
 import '../core/document.dart';
 import '../core/document_editor.dart';
