+++ super_editor/lib/super_editor.dart
@@ -3,6 +3,7 @@ library super_editor;
 export 'package:attributed_text/attributed_text.dart';
 export 'package:super_text_layout/src/caret_layer.dart';
 
+// Fundamental document abstractions
 export 'src/core/document.dart';
 export 'src/core/document_composer.dart';
 export 'src/core/document_debug_paint.dart';
@@ -12,6 +13,8 @@ export 'src/core/document_layout.dart';
 export 'src/core/document_selection.dart';
 export 'src/core/edit_context.dart';
 export 'src/core/styles.dart';
+
+// Super Editor
 export 'src/default_editor/attributions.dart';
 export 'src/default_editor/blockquote.dart';
 export 'src/default_editor/box_component.dart';
@@ -35,12 +38,18 @@ export 'src/default_editor/super_editor.dart';
 export 'src/default_editor/text.dart';
 export 'src/default_editor/text_tools.dart';
 export 'src/default_editor/unknown_component.dart';
+
+// Document operations used by SuperEditor and/or SuperReader,
+// also made available for public use.
+export 'src/document_operations/selection_operations.dart';
+
 export 'src/infrastructure/_listenable_builder.dart';
 export 'src/infrastructure/_logging.dart';
 export 'src/infrastructure/attributed_text_styles.dart';
 export 'src/infrastructure/composable_text.dart';
 export 'src/infrastructure/focus.dart';
 export 'src/infrastructure/ime_input_owner.dart';
+export 'src/infrastructure/keyboard.dart';
 export 'src/infrastructure/multi_tap_gesture.dart';
 export 'src/infrastructure/platforms/android/android_document_controls.dart';
 export 'src/infrastructure/platforms/ios/ios_document_controls.dart';
@@ -48,4 +57,11 @@ export 'src/infrastructure/platforms/mobile_documents.dart';
 export 'src/infrastructure/scrolling_diagnostics/scrolling_diagnostics.dart';
 export 'src/infrastructure/strings.dart';
 export 'src/infrastructure/super_textfield/super_textfield.dart';
+
+// Super Reader
+export 'src/super_reader/read_only_document_android_touch_interactor.dart';
+export 'src/super_reader/read_only_document_ios_touch_interactor.dart';
+export 'src/super_reader/read_only_document_keyboard_interactor.dart';
+export 'src/super_reader/read_only_document_mouse_interactor.dart';
+export 'src/super_reader/reader_context.dart';
 export 'src/super_reader/super_reader.dart';
